+++ packages/language-service/src/utils/ts_utils.ts
