+++ packages/language-service/test/completions_spec.ts
@@ -9,7 +9,10 @@
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
 import ts from 'typescript';
 
-import {DisplayInfoKind, unsafeCastDisplayInfoKindToScriptElementKind} from '../src/display_parts';
+import {
+  DisplayInfoKind,
+  unsafeCastDisplayInfoKindToScriptElementKind,
+} from '../src/utils/display_parts';
 import {LanguageServiceTestEnv, OpenBuffer} from '../testing';
 
 const DIR_WITH_INPUT = {
